requirements.txt
fastapi>=0.110.0,<0.120.0
uvicorn>=0.29.0,<0.30.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.14.3,<2.0.0
crewai>=0.11.2,<0.12.0
instructor>=0.5.2,<0.6.0
pydantic>=2.7.1,<3.0.0
langchain>=0.1.16,<0.2.0
langsmith>=0.1.147,<0.2.0
pydantic-settings>=2.8.1,<3.0.0

